Draft: Split Block from Draft.py
Draft: completed splitting of make block
