+++ src/Mod/Draft/draftmake/make_block.py
@@ -1,0 +1,59 @@
+# ***************************************************************************
+# *   Copyright (c) 2009, 2010 Yorik van Havre <yorik@uncreated.net>        *
+# *   Copyright (c) 2009, 2010 Ken Cline <cline@frii.com>                   *
+# *                                                                         *
+# *   This program is free software; you can redistribute it and/or modify  *
+# *   it under the terms of the GNU Lesser General Public License (LGPL)    *
+# *   as published by the Free Software Foundation; either version 2 of     *
+# *   the License, or (at your option) any later version.                   *
+# *   for detail see the LICENCE text file.                                 *
+# *                                                                         *
+# *   This program is distributed in the hope that it will be useful,       *
+# *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+# *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+# *   GNU Library General Public License for more details.                  *
+# *                                                                         *
+# *   You should have received a copy of the GNU Library General Public     *
+# *   License along with this program; if not, write to the Free Software   *
+# *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
+# *   USA                                                                   *
+# *                                                                         *
+# ***************************************************************************
+"""This module provides the code for Draft make_block function.
+"""
+## @package make_block
+# \ingroup DRAFT
+# \brief This module provides the code for Draft make_block function.
+
+import FreeCAD as App
+
+from draftutils.gui_utils import select
+
+from draftobjects.block import Block
+if App.GuiUp:
+    from draftviewproviders.view_base import ViewProviderDraftPart
+
+
+def make_block(objectslist):
+    """make_block(objectslist)
+    
+    Creates a Draft Block from the given objects.
+
+    Parameters
+    ----------
+    objectlist : list
+        Major radius of the ellipse.
+
+    """
+    if not App.ActiveDocument:
+        App.Console.PrintError("No active document. Aborting\n")
+        return
+    obj = App.ActiveDocument.addObject("Part::Part2DObjectPython","Block")
+    Block(obj)
+    obj.Components = objectslist
+    if App.GuiUp:
+        ViewProviderDraftPart(obj.ViewObject)
+        for o in objectslist:
+            o.ViewObject.Visibility = False
+        select(obj)
+    return obj
