+++ src/Mod/Draft/draftobjects/block.py
@@ -1,0 +1,54 @@
+# ***************************************************************************
+# *   Copyright (c) 2009, 2010 Yorik van Havre <yorik@uncreated.net>        *
+# *   Copyright (c) 2009, 2010 Ken Cline <cline@frii.com>                   *
+# *   Copyright (c) 2019 Eliud Cabrera Castillo <e.cabrera-castillo@tum.de> *
+# *                                                                         *
+# *   This program is free software; you can redistribute it and/or modify  *
+# *   it under the terms of the GNU Lesser General Public License (LGPL)    *
+# *   as published by the Free Software Foundation; either version 2 of     *
+# *   the License, or (at your option) any later version.                   *
+# *   for detail see the LICENCE text file.                                 *
+# *                                                                         *
+# *   This program is distributed in the hope that it will be useful,       *
+# *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+# *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+# *   GNU Library General Public License for more details.                  *
+# *                                                                         *
+# *   You should have received a copy of the GNU Library General Public     *
+# *   License along with this program; if not, write to the Free Software   *
+# *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
+# *   USA                                                                   *
+# *                                                                         *
+# ***************************************************************************
+"""This module provides the object code for Draft Block.
+"""
+## @package block
+# \ingroup DRAFT
+# \brief This module provides the object code for Draft Block.
+
+from PySide.QtCore import QT_TRANSLATE_NOOP
+
+from draftobjects.base import DraftObject
+
+
+class Block(DraftObject):
+    """The Block object"""
+
+    def __init__(self, obj):
+        super(Block, self).__init__(obj, "Block")
+
+        _tip = "The components of this block"
+        obj.addProperty("App::PropertyLinkList","Components",
+                        "Draft",QT_TRANSLATE_NOOP("App::Property", _tip))
+
+    def execute(self, obj):
+        import Part
+        plm = obj.Placement
+        shps = []
+        for c in obj.Components:
+            shps.append(c.Shape)
+        if shps:
+            shape = Part.makeCompound(shps)
+            obj.Shape = shape
+        obj.Placement = plm
+        obj.positionBySupport()
